+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add additional space before the opening bracket for the collections initialization: `new Map<> {`
 - Add option to format standard Apex types references (`apexFormatStandardTypes`)
 - Add option to format properties so that if getter or setter has only one statement they are aligned in one line (`apexExpandOneLineProperties`)
+- Add option to explicitly put the `private` keyword if the access modifier was not provided (`apexExplicitAccessModifier`)
 
 ## Formatting Changes